tport.c: fixed some 64-bit issues.
darcs-hash:20060925192135-65a35-7620f560dd3957017f11bc5c8fa01098def8fcca.gz

+++ libsofia-sip-ua/tport/tport_internal.h
@@ -396,7 +396,7 @@ void tport_hup_event(tport_t *self);
 
 ssize_t tport_recv_iovec(tport_t const *self, 
 			 msg_t **mmsg,
-			 msg_iovec_t iovec[msg_n_fragments], int N, 
+			 msg_iovec_t iovec[msg_n_fragments], size_t N, 
 			 int exact);
 
 msg_t *tport_msg_alloc(tport_t const *self, unsigned size);
